Assert that ShadowNode measurements are within bounds
Summary:
Android's TextLayoutManager may return widths greather than the max measure constraint.

Yoga will clamp these, but this sort of issue points to a logic bug, and creates issues when we are looking at caching text measurements based on constraint reuse.

Let's debug assert that we don't do that, and fix a case of rounding up at a pixel boundary, to ensure that it doesn't go above max width. This should be perfectly safe, since Yoga is already doing this clamping, which is what dictates final size of the TextView.

Changelog: [Internal]

Differential Revision: D74291373