[Fabric} REACT_NATIVE_HAS_SERIALIZABLE_STATE
Summary:
Was running into issues with split Android ParagraphState, because a lot of code is tied to `#ifdef ANDROID`, which also builds for react-native-cxx.

This does that spliitting, and also introduces `REACT_NATIVE_HAS_SERIALIZABLE_STATE` to replace previous ANDROID checks related to state serialization, so we can cofine that to just the Android platform. These changes are dependent on each other.

Changelog: [Internal]

Differential Revision: D74374376