[android][kotlin] make removeView open in ViewGroupManager
After conversion to Kotlin we could no longer override the removeView function since it is no longer open. The rest of this class can be overridden as before, but since functions are final by default this doesn't work for the new `removeView` function.

This commit fixes this by making the removeView function explicitly open.