[kotlin][android] adding accessor for backing map
Expo is using the backingmap to be able to update properties in a fast and performant way in two overridden view manager subclasses (ViewGroupManager subclass is one of them). This was previously done by exposing a helper method accessing the inner field in java due to the inner field being accessible from the same package. This no longer works in 0.80.

This commit fixes this by exposing a public function that returns the backingMap directly.